feat: 新增 algolia 的 docsearch 支持
