+++ packages/preset-dumi/src/themes/default/SearchBar.less
@@ -5,7 +5,7 @@
   position: relative;
   display: inline-block;
 
-  > input {
+  &-input {
     width: 200px;
     height: 32px;
     padding: 0 38px 0 14px;
